feat: 상수 파일에서 타입 분리
